requirements.txt
absl-py                      ==1.2.0
asgiref                      ==3.2.10
astroid                      ==2.12.11
asttokens                    ==2.0.8
astunparse                   ==1.6.3
autopep8                     ==1.7.0
backcall                     ==0.2.0
beautifulsoup4               ==4.11.1
black                        ==22.10.0
blis                         ==0.7.9
bs4                          ===0.0.1
cachetools                   ==5.2.0
catalogue                    ==2.0.8
certifi                      ==2022.9.24
charset-normalizer           ==2.1.1
click                        ==8.1.3
colorama                     ==0.4.5
confection                   ==0.0.3
contourpy                    ==1.0.5
coverage                     ==6.5.0
cycler                       ==0.11.0
cymem                        ==2.0.7
Cython                       ==0.29.28
debugpy                      ==1.6.3
decorator                    ==5.1.1
dill                         ==0.3.5.1
distlib                      ==0.3.6
Django                       ==3.1
djangorestframework          ==3.11.1
docformatter                 ==1.5.0
en-core-web-lg               ==3.4.0
entrypoints                  ==0.4
et-xmlfile                   ==1.1.0
executing                    ==1.1.0
filelock                     ==3.8.0
flake8                       ==5.0.4
flatbuffers                  ==2.0.7
fonttools                    ==4.37.3
gast                         ==0.4.0
gensim                       ==4.2.0
google-auth                  ==2.11.1
google-auth-oauthlib         ==0.4.6
google-pasta                 ==0.2.0
grpcio                       ==1.49.1
h5py                         ==3.7.0
idna                         ==3.4
ipykernel                    ==6.16.0
ipython                      ==8.5.0
isort                        ==5.10.1
jedi                         ==0.18.1
Jinja2                       ==3.1.2
joblib                       ==1.2.0
json2html                    ==1.3.0
jupyter_client               ==7.3.5
jupyter-core                 ==4.11.1
keras                        ==2.10.0
Keras-Preprocessing          ==1.1.2
kiwisolver                   ==1.4.4
langcodes                    ==3.3.0
lazy-object-proxy            ==1.7.1
libclang                     ==14.0.6
libcst                       ==0.4.7
lxml                         ==4.9.1
Markdown                     ==3.4.1
MarkupSafe                   ==2.1.1
matplotlib                   ==3.6.0
matplotlib-inline            ==0.1.6
mccabe                       ==0.7.0
murmurhash                   ==1.0.9
mypy                         ==0.981
mypy-extensions              ==0.4.3
nest-asyncio                 ==1.5.5
nltk                         ==3.7
numpy                        ==1.23.3
oauthlib                     ==3.2.1
openpyxl                     ==3.0.10
opt-einsum                   ==3.3.0
packaging                    ==21.3
pandas                       ==1.5.0
parso                        ==0.8.3
pathspec                     ==0.9.0
pathy                        ==0.6.2
pep8                         ==1.7.1
pep8radius                   ==0.9.2
pickleshare                  ==0.7.5
Pillow                       ==9.2.0
pip                          ==22.2.2
platformdirs                 ==2.5.2
preshed                      ==3.0.8
prompt-toolkit               ==3.0.31
protobuf                     ==3.19.5
psutil                       ==5.9.2
psycopg2                     ==2.9.3
pure-eval                    ==0.2.2
pyasn1                       ==0.4.8
pyasn1-modules               ==0.2.8
pycln                        ==2.1.1
pycodestyle                  ==2.9.1
pydantic                     ==1.9.2
pyflakes                     ==2.5.0
Pygments                     ==2.13.0
pylint                       ==2.15.4
pyparsing                    ==3.0.9
python-dateutil              ==2.8.2
python-dotenv                ==0.21.0
pytz                         ==2022.2.1
pywin32                      ==304
PyYAML                       ==6.0
pyzmq                        ==24.0.1
refurb                       ==1.4.0
regex                        ==2022.9.13
requests                     ==2.28.1
requests-oauthlib            ==1.3.1
rsa                          ==4.9
scikit-learn                 ==1.1.2
scipy                        ==1.9.1
setuptools                   ==65.3.0
six                          ==1.16.0
sklearn                      ==0.0
smart-open                   ==5.2.1
soupsieve                    ==2.3.2.post1
spacy                        ==3.4.1
spacy-legacy                 ==3.0.10
spacy-loggers                ==1.0.3
sqlparse                     ==0.4.2
srsly                        ==2.4.4
stack-data                   ==0.5.1
tensorboard                  ==2.10.1
tensorboard-data-server      ==0.6.1
tensorboard-plugin-wit       ==1.8.1
tensorflow                   ==2.10.0
tensorflow-estimator         ==2.10.0
tensorflow-hub               ==0.12.0
tensorflow-io-gcs-filesystem ==0.27.0
termcolor                    ==2.0.1
textblob                     ==0.17.1
thinc                        ==8.1.4
threadpoolctl                ==3.1.0
toml                         ==0.10.2
tomli                        ==2.0.1
tomlkit                      ==0.11.5
tornado                      ==6.2
tqdm                         ==4.64.1
traitlets                    ==5.4.0
typer                        ==0.4.2
typing_extensions            ==4.3.0
typing-inspect               ==0.8.0
tzdata                       ==2022.2
untokenize                   ==0.1.1
urllib3                      ==1.26.12
virtualenv                   ==20.16.5
wasabi                       ==0.10.1
wcwidth                      ==0.2.5
Werkzeug                     ==2.2.2
wheel                        ==0.37.1
wrapt                        ==1.14.1
xgboost                      ==1.6.2
yapf                         ==0.32.0
